Remove staking target requirement (#18)
This PR is counterpart to
https://github.com/coinbase/staking-client-library-ts/pull/20

It helps remove the requirement to provide an explicit staking target
(integrator contract address for ethereum partial staking and validator
address for sol) while making staking api calls.

As we move to a self-service world, we want customers to be able to just
come and stake to default staking targets i.e. default integration
contracts on both Holesky and Mainnet in the case of Partial ETH staking
and default coinbase validators on both Devnet and Mainnet in the case
of sol staking.

As a result, we have updated the examples to not take integrator
contract or validator address as required inputs.